Properly handle time default in capabilities view
Ugh. Already screwed this up. No default attribute
for individual layers, only for parent layer.

We made an arbitrary choice for the default to always
be the first time in the layer.

Refactored to remove those time related helper methods
as the logic ended up too specific to each use case
to be worth abstracting.