Refactor GT Server for improved temporal handling
OgcSource now properly has a timeInterval as
part of its description since every OGC layer can represent
a time interval as part of its spec.

GeoTrellisRasterSource is handled with a separate layer type
because its the only layer type that supports temporal arguments.
In the future if we add temporal support its possible that
this type could be compressed down into RasterOgcSource.

Construction of the actual raster source responsible for reading
a GeoTrellis layer is now deferred until just before the read.
We no longer construct a list of GeoTrellisRasterSource for all
the valid times in the layer. We only construct one for the
specific time we plan to read based on the request params.

Removes all extra build logic from GeoTrellisRasterSourceLegacy
to facilitate the reversion to upstream GeoTrellisRasterSource.
We'll still need to add the new apply methods, the time param
in the constructor and the time methods. It makes sense for
GeoTrellisRasterSource to request a time parameter because we
need to choose a specific time to read if we want a Tile.

The application.conf and application-spacetimekey.conf were
updated with breaking changes to `layer.layer-name.sources`.

Changes explored but not actually completed here, they were
reverted to simplify diff/merge conflicts:

Rename OgcSource -> OgcLayer to better match OGC terminology

Rename OgcLayer -> OgcResponse to better indicate that this
trait represents the realization of an OGC Layer + OGC Params
into a specific Ogc Response for WMS GetMap/WCS GetCoverage/etc.

Rename all subclasses of these two types to have consistent
postfixes. For example, `SimpleSource extends OgcSource` is
now `SimpleOgcLayer extends OgcLayer`.